requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
av==11.0.0
base58==2.1.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
comm==0.2.1
comtypes==1.4.2
contourpy==1.2.0
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.5
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
entrypoints==0.4
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
ifaddr==0.2.0
importlib-metadata==6.11.0
ipykernel==6.29.2
ipython==8.21.0
jax==0.4.26
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-mistralai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.56
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
mistralai==0.1.8
ml-dtypes==0.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.27.0
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==0.85.0
streamlit-float==0.3.4
streamlit-webrtc==0.47.6
tenacity==8.3.0
threadpoolctl==3.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.40.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
